+++ tests/testthat/test-paired_test_categorical.R
@@ -64,3 +64,10 @@ test_that("matrices are converted to table with a warning", {
   mat_test <- matrix(1:9, nrow = 3, dimnames = list(a = 1:3, b = 1:3))
   expect_warning(paired_test_categorical(mat_test))
 })
+
+
+test_that("input without names do not throw an error", {
+  expect_success(
+   expect_warning(paired_test_categorical(matrix(1:9, nrow = 3)))
+  )
+})